requirements.txt
aiofiles==23.1.0
aiogram==3.0.0
aiohttp==3.8.5
aiosignal==1.3.1
alembic==1.12.0
annotated-types==0.5.0
anyio==3.7.1
apache-libcloud==3.8.0
async-timeout==4.0.3
asyncpg==0.28.0
attrs==23.1.0
Babel==2.13.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
fastapi==0.103.1
fasteners==0.19
frozenlist==1.4.0
greenlet==2.0.2
h11==0.14.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
magic-filter==1.0.11
Mako==1.2.4
MarkupSafe==2.1.3
multidict==6.0.4
Pillow==10.0.1
pydantic==2.3.0
pydantic-settings==2.0.3
pydantic_core==2.6.3
python-dotenv==1.0.0
python-multipart==0.0.6
requests==2.31.0
sniffio==1.3.0
SQLAlchemy==2.0.21
sqlalchemy-file==0.6.0
starlette==0.27.0
starlette-admin==0.11.2
structlog==23.1.0
typing_extensions==4.7.1
urllib3==2.0.7
uvicorn==0.23.2
yarl==1.9.2
